third_party/WebKit/LayoutTests/xr/xrDevice_requestSession_exclusive_supported.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<script src="../xr/resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>


xr_session_promise_test( (session) => {
  assert_not_equals(session, null);
}, { supportsImmersive:true }, [{ immersive: true }],
"requestSession for an immersive session with a user gesture resolves");
</script>
